+++ copier/__init__.py
@@ -2,5 +2,4 @@
 """
 from .main import *  # noqa
 
-
 __version__ = '2.0.1'